feat: update llm-zoo to Zod v4, single source of truth
- Update llm-zoo peerDependencies to zod@^4.0.0
- Update schemas to use z.nativeEnum() for native enums
- TeXRA ModelConfig.ts re-exports schemas from llm-zoo/schemas
- Update README to document Zod v4 requirement
